requirements.txt
astroid==2.11.7
colorama==0.4.5
dependency-injector==4.39.1
dill==0.3.5.1
Faker==13.15.0
greenlet==1.1.2
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mysqlclient==2.1.1
platformdirs==2.5.2
pylint==2.14.4
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.39
tomli==2.0.1
tomlkit==0.11.1
wrapt==1.14.1
